requirements.txt
django-filter==1.0.4
djangorestframework==3.6.4
Django==1.9.8
djangorestframework-jwt==1.11.0
